\name{Maeswrap-package}
\alias{Maeswrap-package}
\alias{Maeswrap}
\docType{package}
\title{Bundle of functions for modifying MAESTRA/MAESPA input files, and reading output files.}
\description{
The main functions are \code{\link{runmaespa}} and \code{\link{maesparunall}}, see their help pages.
Functions that read parameters, and modify parameters or whole namelists are \code{\link{readPAR}}, \code{\link{replacePAR}} and \code{\link{replaceNAMELIST}}. 
Functions that read output are \code{\link{readdayflux}}, \code{\link{readhrflux}}
}
\details{
\tabular{ll}{
Package: \tab Maeswrap\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2008-12-03\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\references{See Belinda Medlyn's MAESTRA homepage at: \url{http://www.bio.mq.edu.au/maestra/}}

\author{
Remko Duursma
Maintainer: Remko Duursma <remkoduursma@gmail.com>
}
\keyword{ package }
