\name{maeswrapdefinitions}
\docType{data}
\alias{maeswrapdefinitions}
\title{Example Maeswrap definition file}
\description{
The MAESTRA/MAESPA wrapper needs a 'definition file', where names of parameters are defined,
together with their locations in parameter files and namelists. The 'comment' column is ignored
when parsing the definition file.
}
\usage{maeswrapdefinitions}
\format{A white space separated dataset (readable with \code{read.table}).}
\references{None.}
\keyword{datasets}