% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmet.R
\name{readmet}
\alias{readmet}
\title{Reads the met.dat input file}
\usage{
readmet(filename = "met.dat", nlines = -1)
}
\arguments{
\item{filename}{Default name of the met.dat file.}

\item{nlines}{Optional, how many lines of the metfile to read?}
}
\value{
Returns a dataframe.
}
\description{
Reads the meteorological input data in the met.dat file.
}
\examples{


\dontrun{

# Simple as pi:
metdata <- readmet()

}

}
\author{
Remko Duursma
}
\keyword{utilities}
