% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{readNameList}
\alias{readNameList}
\title{Read a namelist into a list}
\usage{
readNameList(datfile, namelist)
}
\arguments{
\item{datfile}{Name of the Maestra/Maespa input file (e.g. "trees.dat")}

\item{namelist}{Name of the namelist (e.g. "plot")}
}
\description{
Read a namelist into a list
}

