% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{readuspar}
\alias{readuspar}
\title{Read understorey simulations}
\usage{
readuspar(filename = "uspar.dat")
}
\arguments{
\item{filename}{The understorey file}
}
\description{
Reads the point-wise output file when the understorey was simulated.
}

