% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwatbal.R
\name{readwatbalday}
\alias{readwatbalday}
\title{Reads the watbalday.dat MAESPA output file}
\usage{
readwatbalday(filename = "watbalday.dat")
}
\arguments{
\item{filename}{Default name of the daily water balance output file.}
}
\value{
Returns a dataframe.
}
\description{
Reads the daily water balance output file ("watbalday.dat").
}
\examples{

\dontrun{

# Simple as this:
mywatbalresult <- readwatbal()

# If you want to select the water balance file with a menu:
readwatbalday(file.choose())

}
}
\author{
Rémi Vezy
}
\keyword{utilities}
