\name{runfiletest}
\docType{data}
\alias{runfiletest}
\title{Example Maeswrap run datafile.}
\description{
Example of a 'run dataset', where each row corresponds to a simulation with parameters
set by the column names. The wrapper looks for the column names in the definition file,
and sets the parameters based on information in that file. 
}
\usage{runfiletest}
\format{A comma-separated file.}
\references{None.}
\keyword{datasets}