% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Maeswrap-package.R
\docType{package}
\name{Maeswrap-package}
\alias{Maeswrap}
\alias{Maeswrap-package}
\title{Bundle of functions for modifying MAESTRA/MAESPA input files, and reading
output files.}
\description{
The main functions are \code{\link{runmaespa}} and
\code{\link{maesparunall}}, see their help pages. Functions that read
parameters, and modify parameters or whole namelists are
\code{\link{readPAR}}, \code{\link{replacePAR}},
\code{\link{replaceNameList}}, and \code{\link{parseFile}} to read an entire
file. Functions that read output are \code{\link{readdayflux}},
\code{\link{readhrflux}}
}
\details{
\tabular{ll}{ Package: \tab Maeswrap\cr Type: \tab Package\cr Version: \tab
1.2\cr Date: \tab 2008-12-03\cr License: \tab GPL\cr LazyLoad: \tab yes\cr }
}
\author{
Remko Duursma Maintainer: Remko Duursma <remkoduursma@gmail.com>
}
\references{
See Belinda Medlyn's MAESTRA homepage at:
\url{http://www.bio.mq.edu.au/maestra/}
}
\keyword{package}

