% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maeswrap-package.R
\docType{package}
\name{Maeswrap-package}
\alias{Maeswrap-package}
\alias{Maeswrap}
\title{Bundle of functions for modifying MAESTRA/MAESPA input files, and reading
output files.}
\description{
The main functions are \code{\link[=runmaespa]{runmaespa()}} and
\code{\link[=maesparunall]{maesparunall()}}, see their help pages. Functions that read
parameters, and modify parameters or whole namelists are
\code{\link[=readPAR]{readPAR()}}, \code{\link[=replacePAR]{replacePAR()}},
\code{\link[=replaceNameList]{replaceNameList()}}, and \code{\link[=parseFile]{parseFile()}} to read an entire
file. Functions that read output are \code{\link[=readdayflux]{readdayflux()}},
\code{\link[=readhrflux]{readhrflux()}}
}
\details{
\tabular{ll}{ Package: \tab Maeswrap\cr Type: \tab Package\cr Version: \tab
1.2\cr Date: \tab 2008-12-03\cr License: \tab GPL\cr LazyLoad: \tab yes\cr }
}
\references{
See Belinda Medlyn's MAESTRA homepage at:
\url{http://www.bio.mq.edu.au/maestra/}
}
\author{
Remko Duursma Maintainer: Remko Duursma \href{mailto:remkoduursma@gmail.com}{remkoduursma@gmail.com}
}
\keyword{package}
