\docType{data}
\name{maeswrapdefinitions}
\alias{maeswrapdefinitions}
\title{Example Maeswrap definition file}
\format{A white space separated dataset (readable with \code{read.table}).}
\description{
The MAESTRA/MAESPA wrapper needs a 'definition file', where
names of parameters are defined, together with their
locations in parameter files and namelists. The 'comment'
column is ignored when parsing the definition file.
}
\references{
None.
}
\keyword{datasets}

