% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{readdayflux}
\alias{readdayflux}
\title{Reads the dayflx.dat output file}
\usage{
readdayflux(filename = "dayflx.dat")
}
\arguments{
\item{filename}{Default name of the daily flux file.}
}
\value{
Returns a dataframe.
}
\description{
Reads the dayflx.dat MAESTRA/MAESPA output file, returns a clean dataframe.
Names of the variables are read from the Columns: line.
}
\examples{
\dontrun{
# Read it:
mysim1 <- readdayflux()
}
}
\author{
Remko Duursma
}
\keyword{utilities}

