% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readuspar.R
\name{readuspar}
\alias{readuspar}
\title{Read understorey simulations}
\usage{
readuspar(filename = "uspar.dat")
}
\arguments{
\item{filename}{The understorey file}
}
\description{
Reads the point-wise output file when the understorey was simulated.
}

