% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Maeswrap-package.R
\docType{data}
\name{runfiletest}
\alias{runfiletest}
\title{Example Maeswrap run datafile.}
\format{A comma-separated file.}
\description{
Example of a 'run dataset', where each row corresponds to a simulation with
parameters set by the column names. The wrapper looks for the column names
in the definition file, and sets the parameters based on information in that
file.
}
\references{
None.
}
\keyword{datasets}

